Changes made in structure of registration API
