+++ app.js
@@ -1,31 +1,31 @@
-const createError = require('http-errors');
-const express = require('express');
-const path = require('path');
-const cookieParser = require('cookie-parser');
-const logger = require('morgan');
-const exphbs = require('express-handlebars');
-const indexRouter = require('./routes/index');
-const userRouter = require('./modules/user/userRouter');
-const fileOperationRouter = require('./modules/fileoperation/fileOperationRouter');
-const bodyParser = require('body-parser');
-const cors = require('cors');
+const createError = require("http-errors");
+const express = require("express");
+const path = require("path");
+const cookieParser = require("cookie-parser");
+const logger = require("morgan");
+const exphbs = require("express-handlebars");
+const indexRouter = require("./routes/index");
+const userRouter = require("./modules/user/routes");
+const fileOperationRouter = require("./modules/fileoperation/fileOperationRouter");
+const bodyParser = require("body-parser");
+const cors = require("cors");
 const app = express();
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
-app.engine('handlebars', exphbs({defaultLayout: 'main'}));
-app.set('view engine', 'handlebars');
+app.set("views", path.join(__dirname, "views"));
+app.engine("handlebars", exphbs({ defaultLayout: "main" }));
+app.set("view engine", "handlebars");
 
-app.use(logger('dev'));
+app.use(logger("dev"));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(cors());
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(express.static(path.join(__dirname, "public")));
 
-app.use('/', indexRouter);
-app.use('/api/user', userRouter);
-app.use('/api/files', fileOperationRouter);
+app.use("/", indexRouter);
+app.use("/api/user", userRouter);
+app.use("/api/files", fileOperationRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
@@ -36,11 +36,11 @@ app.use(function(req, res, next) {
 app.use(function(err, req, res, next) {
   // set locals, only providing error in development
   res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
+  res.locals.error = req.app.get("env") === "development" ? err : {};
 
   // render the error page
   res.status(err.status || 500);
-  res.render('error');
+  res.render("error");
 });
 
 module.exports = app;
